--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,10 @@
 module sigs.k8s.io/structured-merge-diff
 
 require gopkg.in/yaml.v2 v2.2.1
+
+require (
+	github.com/json-iterator/go v1.1.6
+	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
+	github.com/modern-go/reflect2 v1.0.1 // indirect
+	github.com/stretchr/testify v1.3.0 // indirect
+)