This is not ready to merge yet. I need to clean up / integrate and optimize the code I copied.

This should actually be a drop-in replacement for the current JSON serialization in the main repo...

Initial benchmark follows, I expect to significantly beat this shortly:

```
BenchmarkFieldSet/serialize-20-12                  30000             63898 ns/op           14104 B/op        188 allocs/op
BenchmarkFieldSet/deserialize-20-12                10000            112561 ns/op           21864 B/op        400 allocs/op
BenchmarkFieldSet/serialize-50-12                  10000            143026 ns/op           31207 B/op        513 allocs/op
BenchmarkFieldSet/deserialize-50-12                 5000            251731 ns/op           53827 B/op       1128 allocs/op
BenchmarkFieldSet/serialize-100-12                  5000            408637 ns/op          100048 B/op       1740 allocs/op
BenchmarkFieldSet/deserialize-100-12                2000            718344 ns/op          174620 B/op       3902 allocs/op
BenchmarkFieldSet/serialize-500-12                  1000           2086484 ns/op          503934 B/op       9237 allocs/op
BenchmarkFieldSet/deserialize-500-12                 500           3413733 ns/op          850048 B/op      19521 allocs/op
BenchmarkFieldSet/serialize-1000-12                  500           4234109 ns/op         1098302 B/op      20014 allocs/op
BenchmarkFieldSet/deserialize-1000-12                200           6758817 ns/op         1851899 B/op      42475 allocs/op
```