@lavalamp 2019-07-22
I'm pretty sure I can still cut down on the allocations a ton. Still WIP.

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkFieldSet/serialize-20-12         63898         26510         -58.51%
BenchmarkFieldSet/deserialize-20-12       112561        91106         -19.06%
BenchmarkFieldSet/serialize-50-12         143026        57427         -59.85%
BenchmarkFieldSet/deserialize-50-12       251731        204589        -18.73%
BenchmarkFieldSet/serialize-100-12        408637        169586        -58.50%
BenchmarkFieldSet/deserialize-100-12      718344        581129        -19.10%
BenchmarkFieldSet/serialize-500-12        2086484       905256        -56.61%
BenchmarkFieldSet/deserialize-500-12      3413733       2712487       -20.54%
BenchmarkFieldSet/serialize-1000-12       4234109       1959082       -53.73%
BenchmarkFieldSet/deserialize-1000-12     6758817       5710877       -15.50%

benchmark                                 old allocs     new allocs     delta
BenchmarkFieldSet/serialize-20-12         188            111            -40.96%
BenchmarkFieldSet/deserialize-20-12       400            349            -12.75%
BenchmarkFieldSet/serialize-50-12         513            301            -41.33%
BenchmarkFieldSet/deserialize-50-12       1128           974            -13.65%
BenchmarkFieldSet/serialize-100-12        1740           1029           -40.86%
BenchmarkFieldSet/deserialize-100-12      3902           3416           -12.46%
BenchmarkFieldSet/serialize-500-12        9237           5696           -38.33%
BenchmarkFieldSet/deserialize-500-12      19521          17103          -12.39%
BenchmarkFieldSet/serialize-1000-12       20014          12371          -38.19%
BenchmarkFieldSet/deserialize-1000-12     42475          37159          -12.52%
```