@lavalamp 2019-07-22
This is still doing way more allocations than it should. Now:

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkFieldSet/serialize-20-12         63898         27979         -56.21%
BenchmarkFieldSet/deserialize-20-12       112561        41165         -63.43%
BenchmarkFieldSet/serialize-50-12         143026        60013         -58.04%
BenchmarkFieldSet/deserialize-50-12       251731        94645         -62.40%
BenchmarkFieldSet/serialize-100-12        408637        169823        -58.44%
BenchmarkFieldSet/deserialize-100-12      718344        257874        -64.10%
BenchmarkFieldSet/serialize-500-12        2086484       928576        -55.50%
BenchmarkFieldSet/deserialize-500-12      3413733       1199376       -64.87%
BenchmarkFieldSet/serialize-1000-12       4234109       1852152       -56.26%
BenchmarkFieldSet/deserialize-1000-12     6758817       2360565       -65.07%

benchmark                                 old allocs     new allocs     delta
BenchmarkFieldSet/serialize-20-12         188            111            -40.96%
BenchmarkFieldSet/deserialize-20-12       400            239            -40.25%
BenchmarkFieldSet/serialize-50-12         513            301            -41.33%
BenchmarkFieldSet/deserialize-50-12       1128           673            -40.34%
BenchmarkFieldSet/serialize-100-12        1740           1029           -40.86%
BenchmarkFieldSet/deserialize-100-12      3902           2401           -38.47%
BenchmarkFieldSet/serialize-500-12        9237           5698           -38.31%
BenchmarkFieldSet/deserialize-500-12      19521          12006          -38.50%
BenchmarkFieldSet/serialize-1000-12       20014          12362          -38.23%
BenchmarkFieldSet/deserialize-1000-12     42475          26186          -38.35%
```