@lavalamp 2019-07-23
It'd be easier to get the allocations down further if we changed the format a bit. Right now encoding things to strings (and back) means an allocation for each key. It's not 100% necessary but fixing probably needs some surgery to jsoniter.

It's a bit weird, but we could do that by leveraging JSON arrays instead of objects. Instead of objects that are lists of (key, subobject) pairs, we could do a list of lists-- ({key type: key}, subobject) pairs.

```
[
  [{"k": {"port":443,"protocol":"tcp"}}, "."],
  [{"f": "spec"}, [
    [{"f": "replicas"}, "."]
  ]
]
```

@apelisse @jennybuckley what do you think?