@lavalamp 2019-07-23
OK I have reasonable names and tests now. Does anyone want to look before I squash?