@lavalamp 2019-07-23
In the main repo, doing nothing but switching in the path element serialization functions here, we get

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      2065260       2130809       +3.17%
BenchmarkUpdateNewObject-12     2415312       2348243       -2.78%
BenchmarkRepeatedUpdate-12      967089        945741        -2.21%
BenchmarkSetToFields-12         22058         19687         -10.75%
BenchmarkFieldsToSet-12         15648         8933          -42.91%
BenchmarkYAMLToTyped-12         65407         66945         +2.35%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      7168           7264           +1.34%
BenchmarkUpdateNewObject-12     8396           8510           +1.36%
BenchmarkRepeatedUpdate-12      2887           3024           +4.75%
BenchmarkSetToFields-12         130            130            +0.00%
BenchmarkFieldsToSet-12         91             70             -23.08%
BenchmarkYAMLToTyped-12         152            152            +0.00%
```