@apelisse 2019-07-23
> It's a bit weird, but we could do that by leveraging JSON arrays instead of objects. Instead of objects that are lists of (key, subobject) pairs, we could do a list of lists-- ({key type: key}, subobject) pairs.

When we created the initial API, we did disregard these kind of APIs because they force you to have an `interface{}` somewhere and we were trying to avoid that (though I maybe see a workaround now).

WRT benchmarks, I'm somewhat surprised by these results. The first 3 benchmarks, which are supposed to be the closest to real-life scenarios are now showing more allocations. The 2-3% difference in CPU time are probably irrelevant. And on the other hand, FieldToSet is dropping almost 50% again (same as last PR)...  I don't know what to think.