@lavalamp 2019-07-23
> WRT benchmarks, I'm somewhat surprised by these results. The first 3 benchmarks, which are supposed to be the closest to real-life scenarios are now showing more allocations. The 2-3% difference in CPU time are probably irrelevant. And on the other hand, FieldToSet is dropping almost 50% again (same as last PR)... I don't know what to think.

To make it weirder, when I hook up everything (using a RawExtension), the opposite happens-- 10-15% improvements except for fieldtoset which gets worse... maybe I compared the wrong things.

Anyway, this is squashed and ready to merge. We can improve it more later, it should be in the right shape at least.