@lavalamp 2019-07-24
Pushed missing boilerplate and another optimization.

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkFieldSet/serialize-20-12         63898         27360         -57.18%
BenchmarkFieldSet/deserialize-20-12       112561        35718         -68.27%
BenchmarkFieldSet/serialize-50-12         143026        57494         -59.80%
BenchmarkFieldSet/deserialize-50-12       251731        86494         -65.64%
BenchmarkFieldSet/serialize-100-12        408637        170144        -58.36%
BenchmarkFieldSet/deserialize-100-12      718344        223283        -68.92%
BenchmarkFieldSet/serialize-500-12        2086484       919350        -55.94%
BenchmarkFieldSet/deserialize-500-12      3413733       1084716       -68.22%
BenchmarkFieldSet/serialize-1000-12       4234109       1786525       -57.81%
BenchmarkFieldSet/deserialize-1000-12     6758817       1950813       -71.14%

benchmark                                 old allocs     new allocs     delta
BenchmarkFieldSet/serialize-20-12         188            110            -41.49%
BenchmarkFieldSet/deserialize-20-12       400            216            -46.00%
BenchmarkFieldSet/serialize-50-12         513            300            -41.52%
BenchmarkFieldSet/deserialize-50-12       1128           616            -45.39%
BenchmarkFieldSet/serialize-100-12        1740           1032           -40.69%
BenchmarkFieldSet/deserialize-100-12      3902           2167           -44.46%
BenchmarkFieldSet/serialize-500-12        9237           5701           -38.28%
BenchmarkFieldSet/deserialize-500-12      19521          10770          -44.83%
BenchmarkFieldSet/serialize-1000-12       20014          12365          -38.22%
BenchmarkFieldSet/deserialize-1000-12     42475          23523          -44.62%
```