@lavalamp 2019-07-24
And

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      2230683       1918830       -13.98%
BenchmarkUpdateNewObject-12     2624417       2175278       -17.11%
BenchmarkRepeatedUpdate-12      1031653       774939        -24.88%
BenchmarkSetToFields-12         24778         8860          -64.24%
BenchmarkFieldsToSet-12         33876         10001         -70.48%
BenchmarkYAMLToTyped-12         68377         64342         -5.90%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      8632           6861           -20.52%
BenchmarkUpdateNewObject-12     11210          7947           -29.11%
BenchmarkRepeatedUpdate-12      3057           2332           -23.72%
BenchmarkSetToFields-12         124            42             -66.13%
BenchmarkFieldsToSet-12         147            82             -44.22%
BenchmarkYAMLToTyped-12         150            152            +1.33%
```